requirements.txt
# Core dependencies
httpx==0.27.0
beautifulsoup4==4.12.3
pandas==2.2.2
openpyxl==3.1.5
lxml==5.3.0
python-dotenv==1.0.1

# Supabase client
supabase==2.7.2

# Google Sheets API (for Phase 3)
google-api-python-client==2.134.0
google-auth==2.31.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utilities
requests==2.32.3
fuzzywuzzy==0.18.0
xlrd==2.0.1

# Web Dashboard
flask==3.0.3
gunicorn==22.0.0